Launch Detection: consolidate in single class (use the only existing method, catapult)
