+++ src/modules/fw_pos_control_l1/launchdetection/CMakeLists.txt
@@ -33,5 +33,4 @@
 
 px4_add_library(launchdetection
 	LaunchDetector.cpp
-	CatapultLaunchMethod.cpp
 )
